add updateWorkout and updateExercise

+++ src/main/java/com/abdoatiia542/GraduationProject/repository/workouts/ExerciseRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ExerciseRepository extends JpaRepository<Exercise, Long> {
+public interface ExerciseRepository extends JpaRepository<Exercise, Integer> {
     @Query("SELECT DISTINCT e FROM Exercise e JOIN FETCH e.bodyFocuses bf WHERE LOWER(bf.name) = LOWER(:name)")
      List<Exercise> findByBodyFocuses(@Param("name")String name);
     Optional<Exercise> findById(Integer id);
